requirements.txt
aws-cdk-lib==2.10.0
constructs>=10.0.0,<11.0.0
requests==2.27.1